requirements.txt
alembic==0.8.6
bcrypt==3.1.0
blinker==1.4
cffi==1.7.0
click==6.6
coverage==4.1
dominate==2.2.1
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.6.0
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Migrate==1.8.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
pycparser==2.14
python-editor==1.0.1
ruamel.yaml==0.14.12
six==1.10.0
SQLAlchemy==1.0.14
visitor==0.1.3
Werkzeug==0.11.10
WTForms==2.1
